% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblast_web.R
\name{countdown}
\alias{countdown}
\title{Countdown seconds}
\usage{
countdown(seconds)
}
\arguments{
\item{seconds}{int}
}
\value{
a letter on console
}
\description{
Countdown seconds
}
\examples{
countdown(30)
}
