% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblast_web.R
\name{get_blast_result}
\alias{get_blast_result}
\title{get blast result from web}
\usage{
get_blast_result(seq, program, database, org)
}
\arguments{
\item{seq}{strings}

\item{program}{strings}

\item{database}{strings}

\item{org}{strings}
}
\value{
"data frame"
}
\description{
get blast result from web
}
\examples{
get_blast_result("MLRACQLSGVTAAAQSCLCGKFVLRPLRPCRRYSTSGSSGLTTGKIAGAGLLFVGGGIGGTILYAKWDSHFRESVEKTIPYSDKLFEMVLGPAAY","tblastn","nr","Homo sapeins")
}
