% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblast_web.R
\name{get_des_df}
\alias{get_des_df}
\title{Mutate a column about description base on Genebank ID}
\usage{
get_des_df(df)
}
\arguments{
\item{df}{"data frome"}
}
\value{
"data frome"
}
\description{
Mutate a column about description base on Genebank ID
}
\examples{
get_cds_df(data_frame)
}
