% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblast_web.R
\name{get_title_df}
\alias{get_title_df}
\title{Mutate a column about description base on Genebank ID}
\usage{
get_title_df(df)
}
\arguments{
\item{df}{"data frame"}
}
\value{
"data frame"
}
\description{
Mutate a column about description base on Genebank ID
}
\examples{
get_title_df(data_frame)
}
