% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rblast_web.R
\name{rblast_web}
\alias{rblast_web}
\title{get all results from blast.ncbi.nlm.nih.gov}
\usage{
rblast_web(seq, program, database, org)
}
\arguments{
\item{seq}{strings}

\item{program}{strings}

\item{database}{strings}

\item{org}{strings}
}
\value{
"data frame"
}
\description{
get all results from blast.ncbi.nlm.nih.gov
}
\examples{
test= rblast_web("MLRACQLSGVTAAAQSCLCGKFVLRPLRPCRRYSTSGSSGLTTGKIAGAGLLFVGGGIGGTILYAKWDSHFRESVEKTIPYSDKLFEMVLGPAAYNVPLPKKMMMMMMMMMMMMMMMMMMMMMMMMMMM","tblastn","nr","Homo sapeins")
}
